requirements.txt
Django==2.2.5
django-crispy-forms==1.7.2
django-discover-runner==1.0
django-role-permissions==2.2.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
Pillow==7.0.0
PyJWT==1.7.1
pytz==2019.2
six==1.14.0
sqlparse==0.3.0
